% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEHR_functions.R
\name{connect_database}
\alias{connect_database}
\title{Open connection to SQLite database}
\usage{
connect_database(dbname)
}
\arguments{
\item{dbname}{Name of SQLite database on hard disk (including full file path relative to working directory)}
}
\description{
Open connection to SQLite database
}
