% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{db_query}
\alias{db_query}
\title{Query an RSQLite database.}
\usage{
db_query(
  codelist,
  db.open = NULL,
  db = NULL,
  db.filepath = NULL,
  tab = c("observation", "drugissue", "hes_primary", "death"),
  codelist.vector = NULL
)
}
\arguments{
\item{codelist}{Name of codelist to query the database with.}

\item{db.open}{An open SQLite database connection created using RSQLite::dbConnect, to be queried.}

\item{db}{Name of SQLITE database on hard disk, to be queried.}

\item{db.filepath}{Full filepath to SQLITE database on hard disk, to be queried.}

\item{tab}{Name of table in SQLite database that is to be queried.}

\item{codelist.vector}{Vector of codes to query the database with. This takes precedent over \code{codelist} if both are specified.}
}
\value{
A data.table with observations contained in the specified codelist.
}
\description{
Query an RSQLite database stored on the hard disk for observations with specific codes.
}
\details{
Specifying \code{db} requires a specific underlying directory structure. The SQLite database must be stored in "data/sql/" relative to the working directory.
If the SQLite database is accessed through \code{db}, the connection will be opened and then closed after the query is complete. The same is true if
the database is accessed through \code{db.filepath}. A connection to the SQLite database can also be opened manually using \code{RSQLite::dbConnect}, and then
using the object as input to parameter \code{db.open}. After wards, the connection must be closed manually using \code{RSQLite::dbDisconnect}. If \code{db.open} is specified, this will take precedence over \code{db} or \code{db.filepath}.

Specifying \code{codelist} requires a specific underlying directory structure. The codelist on the hard disk must be stored in "codelists/analysis/" relative
to the working directory, must be a .csv file, and contain a column "medcodeid", "prodcodeid" or "ICD10" depending on the chosen \code{tab}. The codelist can
also be read in manually, and supplied as a character vector to \code{codelist.vector}. If \code{codelist.vector} is defined, this will take precedence over \code{codelist}.
}
