% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_age.R
\name{extract_age}
\alias{extract_age}
\title{Extract age relative to an index date}
\usage{
extract_age(
  cohort,
  varname = NULL,
  indexdt,
  t = NULL,
  t.varname = TRUE,
  out.save.disk = FALSE,
  out.subdir = NULL,
  out.filepath = NULL,
  return.output = FALSE
)
}
\arguments{
\item{cohort}{Cohort to extract age for.}

\item{varname}{Optional name for variable in output dataset.}

\item{indexdt}{Name of variable which defines index date in \code{cohort}.}

\item{t}{Number of days after index date at which to calculate variable.}

\item{t.varname}{Whether to add \code{t} to \code{varname}.}

\item{out.save.disk}{If \code{TRUE} will attempt to save outputted data frame to directory "data/extraction/".}

\item{out.subdir}{Sub-directory of "data/extraction/" to save outputted data frame into.}

\item{out.filepath}{Full filepath and filename to save outputted data frame into.}

\item{return.output}{If \code{TRUE} will return outputted data frame into R workspace.}
}
\value{
A data frame with variable age
}
\description{
Extract age relative to an index date
}
\details{
Specifying \code{db} requires a specific underlying directory structure. The SQLite database must be stored in "data/sql/" relative to the working directory.
If the SQLite database is accessed through \code{db}, the connection will be opened and then closed after the query is complete. The same is true if
the database is accessed through \code{db.filepath}. A connection to the SQLite database can also be opened manually using \code{RSQLite::dbConnect}, and then
using the object as input to parameter \code{db.open}. After wards, the connection must be closed manually using \code{RSQLite::dbDisconnect}. If \code{db.open} is specified, this will take precedence over \code{db} or \code{db.filepath}.

If \code{out.save.disk = TRUE}, the data frame will automatically be written to an .rds file in a subdirectory "data/extraction/" of the working directory.
This directory structure must be created in advance. \code{out.subdir} can be used to specify subdirectories within "data/extraction/". These options will use a default naming convetion. This can be overwritten
using \code{out.filepath} to manually specify the location on the hard disk to save. Alternatively, return the data frame into the R workspace using \code{return.output = TRUE}
and then save onto the hard disk manually.
}
