% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{extract_patients}
\alias{extract_patients}
\title{Create cohort from patient files}
\usage{
extract_patients(filepath, patids = NULL, select = NULL, set = FALSE)
}
\arguments{
\item{filepath}{Path to directory containing .txt files.}

\item{patids}{Patids of patients to retain in the cohort. Character vector. Numeric values should not be used.}

\item{select}{Character vector of column names to select.}

\item{set}{If \code{TRUE} will create a variable called \code{set} which will contain the number that comes after the word 'set' in the file name.}
}
\value{
Data frame with patient information
}
\description{
Create cohort from patient files
}
