% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_smoking.R
\name{extract_smoking}
\alias{extract_smoking}
\title{Extract smoking status prior to index date.}
\usage{
extract_smoking(
  cohort,
  varname = NULL,
  codelist.non,
  codelist.ex,
  codelist.light,
  codelist.mod,
  codelist.heavy,
  codelist.non.vector,
  codelist.ex.vector,
  codelist.light.vector,
  codelist.mod.vector,
  codelist.heavy.vector,
  indexdt,
  t = NULL,
  t.varname = TRUE,
  db.open = NULL,
  db = NULL,
  db.filepath = NULL,
  out.save.disk = FALSE,
  out.subdir = NULL,
  out.filepath = NULL,
  return.output = TRUE
)
}
\arguments{
\item{cohort}{Cohort to extract age for.}

\item{varname}{Optional name for variable in output dataset.}

\item{codelist.non}{Name of codelist (stored on hard disk in "codelists/analysis/") for non-smoker to query the database with.}

\item{codelist.ex}{Name of codelist (stored on hard disk in "codelists/analysis/") for ex-smoker to query the database with.}

\item{codelist.light}{Name of codelist (stored on hard disk in "codelists/analysis/") for light smoker to query the database with.}

\item{codelist.mod}{Name of codelist (stored on hard disk in "codelists/analysis/") for moderate smoker to query the database with.}

\item{codelist.heavy}{Name of codelist (stored on hard disk in "codelists/analysis/") for heavy smoker to query the database with.}

\item{codelist.non.vector}{Vector of codes for non-smoker to query the database with.}

\item{codelist.ex.vector}{Vector of codes for ex-smoker to query the database with.}

\item{codelist.light.vector}{Vector of codes for light smoker to query the database with.}

\item{codelist.mod.vector}{Vector of codes for moderate smoker to query the database with.}

\item{codelist.heavy.vector}{Vector of codes for heavy smoker to query the database with.}

\item{indexdt}{Name of variable which defines index date in \code{cohort}.}

\item{t}{Number of days after index date at which to calculate variable.}

\item{t.varname}{Whether to add \code{t} to \code{varname}.}

\item{db.open}{An open SQLite database connection created using RSQLite::dbConnect, to be queried.}

\item{db}{Name of SQLITE database on hard disk (stored in "data/sql/"), to be queried.}

\item{db.filepath}{Full filepath to SQLITE database on hard disk, to be queried.}

\item{out.save.disk}{If \code{TRUE} will attempt to save outputted data frame to directory "data/extraction/".}

\item{out.subdir}{Sub-directory of "data/extraction/" to save outputted data frame into.}

\item{out.filepath}{Full filepath and filename to save outputted data frame into.}

\item{return.output}{If \code{TRUE} will return outputted data frame into R workspace.}
}
\value{
A data frame with variable smoking status.
}
\description{
Extract smoking status prior to index date.
}
\details{
Specifying \code{db} requires a specific underlying directory structure. The SQLite database must be stored in "data/sql/" relative to the working directory.
If the SQLite database is accessed through \code{db}, the connection will be opened and then closed after the query is complete. The same is true if
the database is accessed through \code{db.filepath}. A connection to the SQLite database can also be opened manually using \code{RSQLite::dbConnect}, and then
using the object as input to parameter \code{db.open}. After wards, the connection must be closed manually using \code{RSQLite::dbDisconnect}. If \code{db.open} is specified, this will take precedence over \code{db} or \code{db.filepath}.

If \code{out.save.disk = TRUE}, the data frame will automatically be written to an .rds file in a subdirectory "data/extraction/" of the working directory.
This directory structure must be created in advance. \code{out.subdir} can be used to specify subdirectories within "data/extraction/". These options will use a default naming convetion. This can be overwritten
using \code{out.filepath} to manually specify the location on the hard disk to save. Alternatively, return the data frame into the R workspace using \code{return.output = TRUE}
and then save onto the hard disk manually.

Specifying the non-vector type codelists requires a specific underlying directory structure. The codelist on the hard disk must be stored in "codelists/analysis/" relative
to the working directory, must be a .csv file, and contain a column "medcodeid", "prodcodeid" or "ICD10" depending on the chosen \code{tab}. The input
to these variables should just be the name of the files (excluding the suffix .csv). The codelists can also be read in manually, and supplied as a
character vector. This option will take precedence over the codelists stored on the hard disk if both are specified.

We take the most recent smoking status record. If an individuals most recent smoking status is a non-smoker,
but they have a history of smoking prior to this, these individuals will be classed as ex-smokers.
}
