% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_test_data_var.R
\name{extract_test_data_var}
\alias{extract_test_data_var}
\title{Extract standard deviation of all test data values over a specified time period relative to an index date.}
\usage{
extract_test_data_var(
  cohort,
  varname = NULL,
  codelist,
  codelist.vector,
  indexdt,
  t = NULL,
  t.varname = TRUE,
  time.prev = 365.25 * 5,
  time.post = 0,
  lower.bound = -Inf,
  upper.bound = Inf,
  db.open = NULL,
  db = NULL,
  db.filepath = NULL,
  out.save.disk = TRUE,
  out.subdir = NULL,
  out.filepath = NULL,
  return.output = FALSE
)
}
\arguments{
\item{cohort}{Cohort of individuals to extract the 'history of' variable for.}

\item{varname}{Name of variable in the outputted data frame.}

\item{codelist}{Name of codelist (stored on hard disk) to query the database with.}

\item{codelist.vector}{Vector of codes to query the database with. This takes precedent over \code{codelist} if both are specified.}

\item{indexdt}{Name of variable in \code{cohort} which specifies the index date. The extracted variable will be calculated relative to this.}

\item{t}{Number of days after \code{indexdt} at which to extract the variable.}

\item{t.varname}{Whether to alter the variable name in the outputted data frame to reflect \code{t}.}

\item{time.prev}{Number of days prior to index date to look for codes.}

\item{time.post}{Number of days after index date to look for codes.}

\item{lower.bound}{Lower bound for returned values.}

\item{upper.bound}{Upper bound for returned values.}

\item{db.open}{An open SQLite database connection created using RSQLite::dbConnect, to be queried.}

\item{db}{Name of SQLITE database on hard disk (stored in "data/sql/"), to be queried.}

\item{db.filepath}{Full filepath to SQLITE database on hard disk, to be queried.}

\item{out.save.disk}{If \code{TRUE} will attempt to save outputted data frame to directory "data/extraction/".}

\item{out.subdir}{Sub-directory of "data/extraction/" to save outputted data frame into.}

\item{out.filepath}{Full filepath and filename to save outputted data frame into.}

\item{return.output}{If \code{TRUE} will return outputted data frame into R workspace.}
}
\value{
A data frame containing standard deviation of test results.
}
\description{
Extract standard deviation of all test data values over a specified time period relative to an index date.
}
\details{
Specifying \code{db} requires a specific underlying directory structure. The SQLite database must be stored in "data/sql/" relative to the working directory.
If the SQLite database is accessed through \code{db}, the connection will be opened and then closed after the query is complete. The same is true if
the database is accessed through \code{db.filepath}. A connection to the SQLite database can also be opened manually using \code{RSQLite::dbConnect}, and then
using the object as input to parameter \code{db.open}. After wards, the connection must be closed manually using \code{RSQLite::dbDisconnect}. If \code{db.open} is specified, this will take precedence over \code{db} or \code{db.filepath}.

If \code{out.save.disk = TRUE}, the data frame will automatically be written to an .rds file in a subdirectory "data/extraction/" of the working directory.
This directory structure must be created in advance. \code{out.subdir} can be used to specify subdirectories within "data/extraction/". These options will use a default naming convetion. This can be overwritten
using \code{out.filepath} to manually specify the location on the hard disk to save. Alternatively, return the data frame into the R workspace using \code{return.output = TRUE}
and then save onto the hard disk manually.

Currently only returns most recent test result. This will be updated to return more than one most recent test result if specified.
}
