% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_time_until.R
\name{extract_time_until}
\alias{extract_time_until}
\title{Extract a 'time until' type variable}
\usage{
extract_time_until(
  cohort,
  varname.time = NULL,
  varname.indicator = NULL,
  codelist = NULL,
  codelist.vector = NULL,
  indexdt,
  censdt,
  censdt.lag = 0,
  t = NULL,
  t.varname = TRUE,
  db.open = NULL,
  db = NULL,
  db.filepath = NULL,
  tab = c("observation", "drugissue", "hes_primary", "death"),
  out.save.disk = FALSE,
  out.subdir = NULL,
  out.filepath = NULL,
  return.output = FALSE
)
}
\arguments{
\item{cohort}{Cohort of individuals to extract the 'history of' variable for.}

\item{varname.time}{Name of time variable in the outputted data frame.}

\item{varname.indicator}{Name of event/censoring indicator in the outputted data frame.}

\item{codelist}{Name of codelist (stored on hard disk) to query the database with.}

\item{codelist.vector}{Vector of codes to query the database with. This takes precedent over \code{codelist} if both are specified.}

\item{indexdt}{Name of variable in \code{cohort} which specifies the index date. The extracted variable will be calculated relative to this.}

\item{censdt}{Name of variable in \code{cohort} which specifies the censoring date.}

\item{censdt.lag}{Number of days after censoring where events will still be considered, to account for delays in recording.}

\item{t}{Number of days after \code{indexdt} at which to extract the variable.}

\item{t.varname}{Whether to alter the variable name in the outputted data frame to reflect \code{t}.}

\item{db.open}{An open SQLite database connection created using RSQLite::dbConnect, to be queried.}

\item{db}{Name of SQLITE database on hard disk (stored in "data/sql/"), to be queried.}

\item{db.filepath}{Full filepath to SQLITE database on hard disk, to be queried.}

\item{tab}{Table name to query in SQLite database.}

\item{out.save.disk}{If \code{TRUE} will attempt to save outputted data frame to directory "data/extraction/".}

\item{out.subdir}{Sub-directory of "data/extraction/" to save outputted data frame into.}

\item{out.filepath}{Full filepath and filename to save outputted data frame into.}

\item{return.output}{If \code{TRUE} will return outputted data frame into R workspace.}
}
\value{
A data frame with variable patid, a variable containing the time until event/censoring, and a variable containing event/censoring indicator.
}
\description{
Query an RSQLite database and a data frame with the time until first code of interest or censoring, and an event/censoring indicator.
}
\details{
Specifying \code{db} requires a specific underlying directory structure. The SQLite database must be stored in "data/sql/" relative to the working directory.
If the SQLite database is accessed through \code{db}, the connection will be opened and then closed after the query is complete. The same is true if
the database is accessed through \code{db.filepath}. A connection to the SQLite database can also be opened manually using \code{RSQLite::dbConnect}, and then
using the object as input to parameter \code{db.open}. After wards, the connection must be closed manually using \code{RSQLite::dbDisconnect}. If \code{db.open} is specified, this will take precedence over \code{db} or \code{db.filepath}.

If \code{out.save.disk = TRUE}, the data frame will automatically be written to an .rds file in a subdirectory "data/extraction/" of the working directory.
This directory structure must be created in advance. \code{out.subdir} can be used to specify subdirectories within "data/extraction/". These options will use a default naming convetion. This can be overwritten
using \code{out.filepath} to manually specify the location on the hard disk to save. Alternatively, return the data frame into the R workspace using \code{return.output = TRUE}
and then save onto the hard disk manually.

Codelists can be specified in two ways. The first is to read the codelist into R as a character vector and then specify through the argument
\code{codelist.vector}. Codelists stored on the hard disk can also be referred to from the \code{codelist} argument, but require a specific underlying directory structure.
The codelist on the hard disk must be stored in a directory called "codelists/analysis/" relative to the working directory. The codelist must be a .csv file, and
contain a column "medcodeid", "prodcodeid" or "ICD10" depending on the input for argument \code{tab}. The input to argument \code{codelist} should just be a character string of
the name of the files (excluding the suffix '.csv'). The \code{codelist.vector} option will take precedence over the \code{codelist} argument if both are specified.

If \code{dtcens.lag > 0}, then the time until the event of interest will be the time until the minimum of the event of interest, and date of censoring.
}
