% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_read_txt.R
\name{extract_txt_death}
\alias{extract_txt_death}
\title{Read in raw ONS death data file}
\usage{
extract_txt_death(filepath, ..., select = NULL)
}
\arguments{
\item{filepath}{File path to raw .txt file}

\item{...}{Arguments to pass onto utils::read.table}

\item{select}{Character vector of variable names to select}
}
\description{
Read in raw ONS death data file
}
