% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_read_txt.R
\name{extract_txt_drug}
\alias{extract_txt_drug}
\title{Read in raw .txt drugissue file}
\usage{
extract_txt_drug(filepath, ..., select = NULL)
}
\arguments{
\item{filepath}{File path to raw .txt file}

\item{...}{Arguments to pass onto utils::read.table}

\item{select}{Character vector of variable names to select}
}
\description{
Read in raw .txt drugissue file
}
