% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_read_txt.R
\name{extract_txt_linkage}
\alias{extract_txt_linkage}
\title{Read in linkage eligibility file}
\usage{
extract_txt_linkage(filepath, ...)
}
\arguments{
\item{filepath}{File path to raw .txt file}

\item{...}{Arguments to pass onto utils::read.table}
}
\description{
Read in linkage eligibility file
}
