% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_read_txt.R
\name{extract_txt_pat}
\alias{extract_txt_pat}
\title{Read in raw .txt patient file}
\usage{
extract_txt_pat(filepath, ..., set = FALSE)
}
\arguments{
\item{filepath}{File path to raw .txt file}

\item{...}{Arguments to pass onto utils::read.table}

\item{set}{If \code{TRUE} will create a variable called \code{set} which will contain the number that comes after the word 'set' in the file name.}
}
\description{
Read in raw .txt patient file
}
