% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{implement_output}
\alias{implement_output}
\title{Internal function to implement saving extracted variable to disk or returning into R workspace.}
\usage{
implement_output(
  variable.dat,
  varname,
  out.save.disk,
  out.subdir,
  out.filepath,
  return.output
)
}
\description{
Will save extracted variable to disk if \code{out.save.disk = TRUE}. Note it relies on correct underlying structure
of directories. Will output extracted variable into R workspace if \code{return.output = TRUE}.
}
