% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{prep_cohort}
\alias{prep_cohort}
\title{Internal function to prepare cohort when extracting variables.}
\usage{
prep_cohort(cohort, indexdt, t, reduce = TRUE)
}
\arguments{
\item{cohort}{Cohort for which variables are to be extracted.}

\item{indexdt}{Name of the variable specified to be the index date.}

\item{t}{Number of days after index date at which to extract data.}

\item{reduce}{Reduce to \code{patid} and \code{indexdt}.}
}
\description{
Changes the name of the variable specified to be the index date to "indexdt".
If \code{t} has been specified (number of days after index date at which to extract data), \code{t} is added to \code{indexdt}.
}
