% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{prep_subdir}
\alias{prep_subdir}
\title{Internal function to prepare subdirectory for saving extracted variables to disk.}
\usage{
prep_subdir(out.subdir)
}
\description{
If out.subdir has been specified, check if it exists. If not, create directory.
}
