% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{prep_varname}
\alias{prep_varname}
\title{Internal function to prepare variable name.}
\usage{
prep_varname(varname, t, t.varname)
}
\description{
Adds the value of \code{t} to variable name is \code{t.varname = TRUE}. May be of use if landmarking and extracting cohort at various intervals post baseline.
}
