% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAURUM-package.R
\docType{package}
\name{rAURUM-package}
\alias{rAURUM}
\alias{rAURUM-package}
\title{rAURUM: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alexpate30.github.io/rAURUM/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
