% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory_systems.R
\name{create_directory_system}
\alias{create_directory_system}
\title{Create the appropriate root directory system to be able to run functions with specifying hard filepaths}
\usage{
create_directory_system(rootdir = NULL)
}
\arguments{
\item{rootdir}{Directory within which to create the directory system}
}
\description{
Create the appropriate root directory system to be able to run functions with specifying hard filepaths
}
